A quick and dirty tool for checking websockets work
Due to strange network configurations we sometimes see users who can't use
jupyter or rstudio/shiny-apps. This is usually because of a problem with their
network setup. This is a service we can deploy and point their IT helpdesk to so
they can see the problem (because they can't login to the platform)